src/popupservice/requirements.txt
grpcio==1.69.0
grpcio-tools==1.69.0
protobuf==5.29.3
